lambda-functions/requirements.txt
requests
bs4
lxml
psycopg2-binary
selenium
pandas
json
